requirements.txt
requests
python-telegram-bot
elasticsearch
logstash
psycopg2-binary
python-dotenv
zstandard
elasticsearch>=8.0.0,<9.0.0
PyYAML
watchdog
telethon
